+++ src/Resources/app/storefront/src/checkout/swag-paypal.abstract-standalone.js
@@ -6,6 +6,9 @@ import SwagPaypalAbstractButtons from '../swag-paypal.abstract-buttons';
 import SwagPayPalScriptLoading from '../swag-paypal.script-loading';
 
 export default class SwagPaypalAbstractStandalone extends SwagPaypalAbstractButtons {
+    /**
+     * @deprecated tag:v10.0.0 - will be removed without replacement
+     */
     static scriptLoading = new SwagPayPalScriptLoading();
     static product = 'spb';
 
@@ -13,53 +16,11 @@ export default class SwagPaypalAbstractStandalone extends SwagPaypalAbstractButt
         ...super.options,
 
         /**
-         * This option holds the client id specified in the settings
-         *
-         * @type string
-         */
-        clientId: '',
-
-        /**
-         * This option holds the merchant id specified in the settings
-         *
-         * @type string
-         */
-        merchantPayerId: '',
-
-        /**
          * This option holds the client token required for field rendering
          *
          * @type string
          */
         clientToken: '',
-
-        /**
-         * This options specifies the currency of the PayPal button
-         *
-         * @type string
-         */
-        currency: 'EUR',
-
-        /**
-         * This options defines the payment intent
-         *
-         * @type string
-         */
-        intent: 'capture',
-
-        /**
-         * This option toggles the PayNow/Login text at PayPal
-         *
-         * @type boolean
-         */
-        commit: true,
-
-        /**
-         * This option specifies the language of the PayPal button
-         *
-         * @type string
-         */
-        languageIso: 'en_GB',
 
         /**
          * This option specifies the PayPal button color