+++ src/Resources/app/storefront/src/swag-paypal.script-loading.js
@@ -1,3 +1,6 @@
+/**
+ * @deprecated tag:v10.0.0 - will be removed without replacement
+ */
 export default class SwagPayPalScriptLoading {
     loadingScript = false;