+++ src/Storefront/Data/Service/AbstractCheckoutDataService.php
@@ -24,7 +24,7 @@ use Swag\PayPal\Util\LocaleCodeProvider;
 use Symfony\Component\Routing\RouterInterface;
 
 #[Package('checkout')]
-abstract class AbstractCheckoutDataService
+abstract class AbstractCheckoutDataService extends AbstractScriptDataService
 {
     public const PAYPAL_ERROR = 'isPayPalError';
 
@@ -33,11 +33,12 @@ abstract class AbstractCheckoutDataService
      */
     public function __construct(
         private readonly PaymentMethodDataRegistry $paymentMethodDataRegistry,
-        private readonly LocaleCodeProvider $localeCodeProvider,
+        LocaleCodeProvider $localeCodeProvider,
         private readonly RouterInterface $router,
-        protected readonly SystemConfigService $systemConfigService,
-        protected readonly CredentialsUtilInterface $credentialsUtil
+        SystemConfigService $systemConfigService,
+        CredentialsUtilInterface $credentialsUtil
     ) {
+        parent::__construct($localeCodeProvider, $systemConfigService, $credentialsUtil);
     }
 
     abstract public function buildCheckoutData(SalesChannelContext $context, ?Cart $cart = null, ?OrderEntity $order = null): ?AbstractCheckoutData;
@@ -49,24 +50,19 @@ abstract class AbstractCheckoutDataService
 
     protected function getBaseData(SalesChannelContext $context, ?OrderEntity $order = null): array
     {
+        if ($context->getCustomer() === null) {
+            throw CartException::customerNotLoggedIn();
+        }
+
         $paymentMethodId = $this->paymentMethodDataRegistry->getEntityIdFromData(
             $this->paymentMethodDataRegistry->getPaymentMethod($this->getMethodDataClass()),
             $context->getContext()
         );
 
         $salesChannelId = $context->getSalesChannelId();
-        $customer = $context->getCustomer();
-
-        if ($customer === null) {
-            throw CartException::customerNotLoggedIn();
-        }
 
         $data = [
-            'clientId' => $this->credentialsUtil->getClientId($salesChannelId),
-            'merchantPayerId' => $this->credentialsUtil->getMerchantPayerId($salesChannelId),
-            'languageIso' => $this->getButtonLanguage($context),
-            'currency' => $context->getCurrency()->getIsoCode(),
-            'intent' => \mb_strtolower($this->systemConfigService->getString(Settings::INTENT, $salesChannelId)),
+            ...parent::getBaseData($context, $order),
             'buttonShape' => $this->systemConfigService->getString(Settings::SPB_BUTTON_SHAPE, $salesChannelId),
             'paymentMethodId' => $paymentMethodId,
             'createOrderUrl' => $context->hasExtension('subscription')
@@ -103,16 +99,5 @@ abstract class AbstractCheckoutDataService
         }
 
         return $data;
-    }
-
-    private function getButtonLanguage(SalesChannelContext $context): string
-    {
-        if ($settingsLocale = $this->systemConfigService->getString(Settings::SPB_BUTTON_LANGUAGE_ISO, $context->getSalesChannelId())) {
-            return $this->localeCodeProvider->getFormattedLocaleCode($settingsLocale);
-        }
-
-        return $this->localeCodeProvider->getFormattedLocaleCode(
-            $this->localeCodeProvider->getLocaleCodeFromContext($context->getContext())
-        );
     }
 }