+++ src/Storefront/Data/Service/FundingEligibilityDataService.php
@@ -12,25 +12,14 @@ use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
 use Swag\PayPal\Checkout\SalesChannel\MethodEligibilityRoute;
 use Swag\PayPal\Setting\Service\CredentialsUtilInterface;
-use Swag\PayPal\Setting\Settings;
 use Swag\PayPal\Storefront\Data\Struct\FundingEligibilityData;
 use Swag\PayPal\Util\LocaleCodeProvider;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Routing\RouterInterface;
 
 #[Package('checkout')]
-class FundingEligibilityDataService
+class FundingEligibilityDataService extends AbstractScriptDataService
 {
-    private CredentialsUtilInterface $credentialsUtil;
-
-    private SystemConfigService $systemConfigService;
-
-    private LocaleCodeProvider $localeCodeProvider;
-
-    private RouterInterface $router;
-
-    private RequestStack $requestStack;
-
     /**
      * @internal
      */
@@ -38,40 +27,19 @@ class FundingEligibilityDataService
         CredentialsUtilInterface $credentialsUtil,
         SystemConfigService $systemConfigService,
         LocaleCodeProvider $localeCodeProvider,
-        RouterInterface $router,
-        RequestStack $requestStack
+        private readonly RouterInterface $router,
+        private readonly RequestStack $requestStack
     ) {
-        $this->credentialsUtil = $credentialsUtil;
-        $this->systemConfigService = $systemConfigService;
-        $this->localeCodeProvider = $localeCodeProvider;
-        $this->router = $router;
-        $this->requestStack = $requestStack;
+        parent::__construct($localeCodeProvider, $systemConfigService, $credentialsUtil);
     }
 
     public function buildData(SalesChannelContext $context): ?FundingEligibilityData
     {
-        return (new FundingEligibilityData())->assign(
-            [
-                'clientId' => $this->credentialsUtil->getClientId($context->getSalesChannelId()),
-                'merchantPayerId' => $this->credentialsUtil->getMerchantPayerId($context->getSalesChannelId()),
-                'languageIso' => $this->getButtonLanguage($context),
-                'currency' => $context->getCurrency()->getIsoCode(),
-                'intent' => \mb_strtolower($this->systemConfigService->getString(Settings::INTENT, $context->getSalesChannelId())),
-                'methodEligibilityUrl' => $this->router->generate('frontend.paypal.payment-method-eligibility'),
-                'filteredPaymentMethods' => $this->getFilteredPaymentMethods(),
-            ]
-        );
-    }
-
-    private function getButtonLanguage(SalesChannelContext $context): string
-    {
-        if ($settingsLocale = $this->systemConfigService->getString(Settings::SPB_BUTTON_LANGUAGE_ISO, $context->getSalesChannelId())) {
-            return $this->localeCodeProvider->getFormattedLocaleCode($settingsLocale);
-        }
-
-        return $this->localeCodeProvider->getFormattedLocaleCode(
-            $this->localeCodeProvider->getLocaleCodeFromContext($context->getContext())
-        );
+        return (new FundingEligibilityData())->assign([
+            ...parent::getBaseData($context),
+            'methodEligibilityUrl' => $this->router->generate('frontend.paypal.payment-method-eligibility'),
+            'filteredPaymentMethods' => $this->getFilteredPaymentMethods(),
+        ]);
     }
 
     private function getFilteredPaymentMethods(): array