+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 8.4.0
+- PPI-807 - Improved script loading performance in the Storefront
 - PPI-951 -  Fixes an issue, where only generic error messages were displayed during checkout without Smart Payment Buttons
 - PPI-953 - Fixes an issue, where shipping tracking codes were not synced with an invalid carrier